+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - Bug #562: Fix `ActiveRecordInterface::upsert()` to prioritize passed associative values during updates (@Tigrov)
 - Bug #561: Fix `ActiveRecordInterface::upsert()` with `$updateProperties = false` (@Tigrov)
 - Bug #550: Relation query should be created by related class, not primary model class (@batyrmastyr)
+- Enh #571: Optimize performance of `ActiveRecord::get()` method (@Tigrov)
+- Enh #575: Remove check for empty string in `AbstractActiveRecord::markPropertyChanged()` method (@Tigrov)
+- Enh #576, #577: Add default config for `yiisoft/config` plugin (@Tigrov)
 
 ## 1.0.2 March 11, 2026